backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Supabase
supabase>=2.10.0

# Google Gemini (embeddings + LLM)
google-generativeai==0.7.2

# LangGraph agent orchestration
langgraph==0.1.5
langchain-core==0.2.0

# Scraping (NO headless browsers)
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0

# PDF parsing
PyPDF2==3.0.1

# Env vars
python-dotenv==1.0.1
